dev-requirements.txt
Flask-Login
Flask-Testing
Flask-SQLAlchemy
nose
nose-progressive
coverage
pyyaml
